feat: Add a CodeSnippet React component
A new CodeSnippet component with copy and paste functionality, when
hovering over the code snippet, a message will be displayed to say that
clicking on the component will copy the code to the clipboard, once
clicked a confirmation message will appear that disappears after a
couple of seconds.

This commit also configures the initial Chrome extension manifest and
Vite build configuration, and configures Tailwind for presentation and
Jest for component/unit testing.