feat: Add a background worker
Adds a background worker that listens for AWS login requests and
intercepts the SAML assertion, before exchanging it for temporary STS
credentials and persisting them to local storage. A limitation of this
pull request is that only a single set of credentials can be persisted,
future versions will be able to persist many credential sets and capable
of removing them automatically once the session duration has expired.